+++ rtl/ibus/ibdlib.vhd
@@ -1,4 +1,4 @@
--- $Id: ibdlib.vhd 1139 2019-04-27 14:00:38Z mueller $
+-- $Id: ibdlib.vhd 1142 2019-04-28 19:27:57Z mueller $
 --
 -- Copyright 2008-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
@@ -19,6 +19,7 @@
 -- Tool versions:  ise 8.2-14.7; viv 2014.4-2018.3; ghdl 0.18-0.35
 -- Revision History: 
 -- Date         Rev Version  Comment
+-- 2019-04-28  1142   1.3.8  add ibd_m9312
 -- 2019-04-26  1139   1.3.7  add ibdr_dl11_buf
 -- 2019-04-14  1131   1.3.6  RLIM_CEV now slv8
 -- 2019-04-07  1129   1.3.5  add ibdr_pc11_buf
@@ -323,6 +324,16 @@ component ibdr_lp11_buf is              -- ibus dev(rem): LP11 (buffered)
     IB_SRES : out ib_sres_type;         -- ibus response
     EI_REQ : out slbit;                 -- interrupt request
     EI_ACK : in slbit                   -- interrupt acknowledge
+  );
+end component;
+
+component ibd_m9312 is                  -- ibus dev: M9312
+                                        -- fixed address: 165***,173***
+  port (
+    CLK : in slbit;                     -- clock
+    RESET : in slbit;                   -- system reset
+    IB_MREQ : in ib_mreq_type;          -- ibus request
+    IB_SRES : out ib_sres_type          -- ibus response
   );
 end component;